# Server/API/Yaml file/openapi.yaml
openapi: 3.0.0
info:
  title: Shay & Gal - Recipes API
  description: This is an uptaded API for Assignment 3.2
  contact:
    name: Shay Eretz Kdosha & Gal Buzaglo
    email: shayek34@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.3.3
servers:
- url: https://virtserver.swaggerhub.com/GalBuzi/assignment3/1.0.0
  description: SwaggerHub API Auto Mocking
- url: https://virtserver.swaggerhub.com/ShayEK34/Example/1.0.0
  description: SwaggerHub API Auto Mocking
paths:
...
  /recipes/getRecipe/id/{recipeId}:
    get:
      tags:
      - Recipes
      summary: Get full recipe info by id.
      description: get information about spesific recipe
      parameters:
      - name: recipeId
        in: path
        description: The ID of the recipe
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 633508
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fullRecipeInfo'
        "404":
          description: No results found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        "500":
          description: Server error occured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
...
  /recipes/getRandomRecipes:
    get:
      tags:
      - Recipes
      summary: return three random recipes
      description: The server return three random recipes.
      responses:
        "200":
          description: Get three random recipes maped by key - recipe id and value
            - object of recipe.
          content:
            application/json:
              schema:
                type: array
                example:
                - id: 29474
                  title: pizza
                  readyInMinutes: 30
                  aggregateLikes: 100
                  vegetrian: true
                  vegan: false
                  glutenFree: false
                  image: https://spoonacular.com/recipeImages/640509-556x370.jpg
                - id: 30154
                  title: burger
                  readyInMinutes: 45
                  aggregateLikes: 200
                  vegetrian: false
                  vegan: false
                  glutenFree: false
                  image: https://spoonacular.com/recipeImages/640509-556x370.jpg
                - id: 16424
                  title: pasta
                  readyInMinutes: 15
                  aggregateLikes: 50
                  vegetrian: true
                  vegan: true
                  glutenFree: false
                  image: recipe.jpghttps://spoonacular.com/recipeImages/640509-556x370.jpg
                items:
                  $ref: '#/components/schemas/recipe'
        "404":
          description: No results found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        "500":
          description: Server error occured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
...
components:
  schemas:
    fullRecipeInfo:
      type: object
      properties:
        id:
          type: integer
          description: The recipe identifier, returned for futher processing and actions
            on recipes.
          example: 633508
        title:
          type: string
          example: Baked Cheese Manicotti
        readyInMinutes:
          type: integer
          description: the recipe time duration
          example: 45
        aggregateLikes:
          type: integer
          example: 1
        vegetrian:
          type: boolean
          example: false
        vegan:
          type: boolean
          example: false
        glutenFree:
          type: boolean
          example: false
        image:
          type: string
          description: the recipe image url
          example: recipe.jpg
        ingredients:
          $ref: '#/components/schemas/recipeIngredients'
        instructions:
          $ref: '#/components/schemas/recipeStepsInstructions'
    recipe:
      type: object
      properties:
        id:
          type: integer
          description: The recipe identifier, returned for futher processing and actions
            on recipes.
          example: 29
        title:
          type: string
          example: pizza
        readyInMinutes:
          type: integer
          description: the recipe time duration
          example: 30
        aggregateLikes:
          type: integer
          example: 100
        vegetrian:
          type: boolean
          example: true
        vegan:
          type: boolean
          example: false
        glutenFree:
          type: boolean
          example: false
        image:
          type: string
          description: the recipe image url
          example: recipe.jpg
...
    recipeStepsInstructions:
      type: array
      example:
      - step_number: "1"
        step: Grease a 9" x 13" baking pan with 1 tbsp. butter and spread 12 cup of
          the marinara sauce across the bottom of the pan. Set aside. Bring a 6-qt.
          pot of salted water to a boil over high heat.
      - step_number: "2"
        step: Add the manicotti and cook until just tender, about 8 minutes.
      - step_number: "3"
        step: Drain manicotti and rinse under cold water; set aside.
      - step_number: "4"
        step: Heat oven to 45
      - step_number: "5"
        step: Heat remaining butter in a 12" skillet over medium heat.
      - step_number: "6"
        step: Add garlic and cook, stirring occasionally, until soft, about 5 minutes.
      - step_number: "7"
        step: Transfer garlic to a medium bowl along with the ricotta, 12 cup parmesan,
          5 tbsp. chopped parsley, salt, pepper, nutmeg, and eggs and stir to combine.Spoon
          some of the filling into both openings of each manicotti shell. (Alternatively,
          transfer the ricotta mixture to a 1-gallon resealable plastic bag, snip
          off a bottom corner of the bag, and pipe filling into pasta.) Repeat with
          remaining manicotti shells.
      - step_number: "8"
        step: Transfer stuffed manicotti to prepared baking dish, making 2 rows.
      - step_number: "9"
        step: Spread the remaining marinara sauce over the manicotti and sprinkle
          with remaining parmesan.
      - step_number: "10"
        step: Bake until hot and bubbly, about 20 minutes. Sprinkle with remaining
          parsley.
      - step_number: "11"
        step: Let sit for 5 minutes before serving.
      items:
        type: string
...
    recipeIngredients:
      type: array
      example:
      - ingredient_name: bell pepper
        amount: 6
        unit: servings
      - ingredient_name: eggs
        amount: 2
        unit: ""
      - ingredient_name: garlic
        amount: 2
        unit: cloves
      - ingredient_name: marinara sauce
        amount: 3
        unit: cups
      - ingredient_name: nutmeg
        amount: 1
        unit: tsp
      - ingredient_name: parmesan
        amount: 1
        unit: cup
      - ingredient_name: parsley
        amount: 2
        unit: tablespoons
      - ingredient_name: shells
        amount: 8
        unit: oz
      - ingredient_name: skim milk ricotta
        amount: 4
        unit: cups
      - ingredient_name: unsalted butter
        amount: 3
        unit: tablespoons
      items:
        type: string
...
    Error404:
      type: object
      properties:
        Msg:
          type: string
          example: Page not found
...
    Error500:
      type: object
      properties:
        Msg:
          type: string
          example: Server error
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      description: Session key to identify user session
      name: session
      in: cookie